# docs/openapi.yml
openapi: 3.0.1
info:
  title: nfcwalker
  version: 1.0.0
paths:
...
  /api/scan/events/{eventId}/photos:
    post:
      operationId: addPhotosToScanEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photos:
                  type: array
                  nullable: true
                  items:
                    type: string
                    format: binary
            encoding:
              photos:
                contentType: application/octet-stream
                explode: false
        required: true
      responses:
        "200":
          description: addPhotosToScanEvent 200 response
          content:
            application/json:
              schema:
                type: object
  /api/scan/events/{eventId}/photos/{photoId}:
    delete:
      operationId: deletePhotoFromScanEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: photoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: deletePhotoFromScanEvent 200 response
  /api/scan/finish:
    post:
      operationId: finishScanMultipart
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - metadata
              type: object
              properties:
                metadata:
                  $ref: "#/components/schemas/FinishScanRequest"
                photos:
                  type: array
                  nullable: true
                  items:
                    type: string
                    format: binary
            encoding:
              metadata:
                contentType: application/octet-stream
                explode: false
              photos:
                contentType: application/octet-stream
                explode: false
          application/json:
            schema:
              $ref: "#/components/schemas/FinishScanRequest"
            example: null
        required: true
      responses:
        "200":
          description: finishScanMultipart 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinishScanResponse"
  /api/scan/start:
    post:
      operationId: startScan
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartScanRequest"
        required: true
      responses:
        "200":
          description: startScan 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartScanResponse"
  /api/sites:
    get:
      operationId: listSites
      parameters:
      - name: organizationId
        in: query
        required: true
        explode: false
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: listSites 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_SiteResponse_"
    post:
      operationId: createSite
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteRequest"
        required: true
      responses:
        "200":
          description: createSite 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteResponse"
  /api/sites/{id}:
    get:
      operationId: getSite
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: getSite 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteResponse"
    put:
      operationId: updateSite
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSiteRequest"
        required: true
      responses:
        "200":
          description: updateSite 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteResponse"
    delete:
      operationId: deleteSite
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: deleteSite 200 response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
...
components:
  schemas:
...
    CheckStatus:
      type: string
      enum:
      - OK
      - PROBLEMS_FOUND
      - SKIPPED
...
    CreateSiteRequest:
      required:
      - name
      - organizationId
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
        name:
          type: string
...
    FinishScanRequest:
      required:
      - challenge
      - scannedAt
      type: object
      properties:
        challenge:
          type: string
        scannedAt:
          type: string
          format: date-time
        lat:
          type: number
          nullable: true
        lon:
          type: number
          nullable: true
        checkStatus:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/CheckStatus"
        checkNotes:
          type: string
          nullable: true
        subCheckResults:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/SubCheckResultRequest"
        incidents:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/IncidentCreateRequest"
    FinishScanResponse:
      required:
      - eventId
      - verdict
      type: object
      properties:
        eventId:
          type: string
          format: uuid
        verdict:
          $ref: "#/components/schemas/ScanVerdict"
    GeoConstraint:
      required:
      - lat
      - lon
      - radiusM
      type: object
      properties:
        lat:
          type: number
        lon:
          type: number
        radiusM:
          type: number
    IncidentCreateRequest:
      required:
      - description
      - severity
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
          nullable: true
        siteId:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
        severity:
          $ref: "#/components/schemas/IncidentSeverity"
        checkpointId:
          type: string
          format: uuid
          nullable: true
        scanEventId:
          type: string
          format: uuid
          nullable: true
...
    IncidentSeverity:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      - CRITICAL
...
    Page_SiteResponse_:
      required:
      - totalPages
      - totalSize
      type: object
      allOf:
      - $ref: "#/components/schemas/Slice_SiteResponse_"
      - type: object
        properties:
          totalSize:
            type: integer
            format: int64
          totalPages:
            type: integer
            format: int32
    Pageable:
      required:
      - mode
      - number
      - size
      - sort
      type: object
      allOf:
      - $ref: "#/components/schemas/Sort"
      - type: object
        properties:
          number:
            type: integer
            format: int32
          size:
            type: integer
            format: int32
          mode:
            $ref: "#/components/schemas/Pageable.Mode"
          sort:
            $ref: "#/components/schemas/Sort"
    Pageable.Mode:
      type: string
      enum:
      - CURSOR_NEXT
      - CURSOR_PREVIOUS
      - OFFSET
...
    ScanPolicy:
      required:
      - checkpointId
      - order
      - runId
      - timeWindow
      type: object
      properties:
        runId:
          type: string
          format: uuid
        checkpointId:
          type: string
          format: uuid
        order:
          type: integer
          format: int32
        timeWindow:
          $ref: "#/components/schemas/TimeWindow"
        geo:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/GeoConstraint"
    ScanVerdict:
      type: string
      enum:
      - OK
      - WARNING
      - FAIL
    SiteResponse:
      required:
      - createdAt
      - id
      - name
      - organizationId
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Slice_SiteResponse_:
      required:
      - content
      - empty
      - numberOfElements
      - offset
      - pageNumber
      - pageable
      - size
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/SiteResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        pageNumber:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        empty:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
    Sort:
      required:
      - orderBy
      type: object
      properties:
        orderBy:
          type: array
          items:
            $ref: "#/components/schemas/Sort.Order"
    Sort.Order:
      required:
      - ascending
      - direction
      - ignoreCase
      - property
      type: object
      properties:
        ignoreCase:
          type: boolean
        direction:
          $ref: "#/components/schemas/Sort.Order.Direction"
        property:
          type: string
        ascending:
          type: boolean
    Sort.Order.Direction:
      type: string
      enum:
      - ASC
      - DESC
    StartScanRequest:
      required:
      - checkpointCode
      - deviceId
      - organizationId
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
        deviceId:
          type: string
        checkpointCode:
          type: string
    StartScanResponse:
      required:
      - challenge
      - policy
      type: object
      properties:
        challenge:
          type: string
        policy:
          $ref: "#/components/schemas/ScanPolicy"
...
    SubCheckResultRequest:
      required:
      - status
      - subCheckId
      type: object
      properties:
        subCheckId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/CheckStatus"
        notes:
          type: string
          nullable: true
    TimeWindow:
      required:
      - maxOffsetSec
      - minOffsetSec
      type: object
      properties:
        minOffsetSec:
          type: integer
          format: int32
        maxOffsetSec:
          type: integer
          format: int32
...
    UpdateSiteRequest:
      required:
      - name
      type: object
      properties:
        name:
          type: string
...