# src/main/kotlin/ge/tiger8bit/controller/SiteController.kt
package ge.tiger8bit.controller

import ge.tiger8bit.dto.CreateSiteRequest
import ge.tiger8bit.dto.SiteResponse
import ge.tiger8bit.dto.UpdateSiteRequest
import ge.tiger8bit.service.SiteService
import io.micronaut.data.model.Page
import io.micronaut.data.model.Pageable
import io.micronaut.http.annotation.*
import io.micronaut.security.annotation.Secured
import java.security.Principal
import java.util.*

@Controller("/api/sites")
@Secured("ROLE_BOSS", "ROLE_APP_OWNER")
class SiteController(
    private val siteService: SiteService
) {
    @Post
    fun createSite(@Body request: CreateSiteRequest, principal: Principal): SiteResponse {
        val userId = UUID.fromString(principal.name)
        return siteService.create(request, userId)
    }

    @Get
    fun listSites(
        @QueryValue organizationId: UUID,
        @QueryValue(defaultValue = "0") page: Int,
        @QueryValue(defaultValue = "20") size: Int,
        principal: Principal
    ): Page<SiteResponse> {
        val userId = UUID.fromString(principal.name)
        val pageable = Pageable.from(page, size)
        return siteService.findByOrganizationId(organizationId, userId, pageable)
    }

    @Get("/{id}")
    fun getSite(@PathVariable id: UUID, principal: Principal): SiteResponse? {
        val userId = UUID.fromString(principal.name)
        return siteService.findById(id, userId)
    }

    @Put("/{id}")
    fun updateSite(
        @PathVariable id: UUID,
        @Body request: UpdateSiteRequest,
        principal: Principal
    ): SiteResponse {
        val userId = UUID.fromString(principal.name)
        return siteService.update(id, request, userId)
    }

    @Delete("/{id}")
    fun deleteSite(@PathVariable id: UUID, principal: Principal): Map<String, Any> {
        val userId = UUID.fromString(principal.name)
        siteService.delete(id, userId)
        return mapOf("deleted" to true, "id" to id)
    }
}